syscall: remove (long) from .c and add __long on syscall_asm.h (only x86_64)

+++ src/syscall/_exit.c
@@ -4,5 +4,5 @@
 void
 _exit(int exit_code)
 {
-	__syscall1(SYS_exit, (long)exit_code);
+	__syscall1(SYS_exit, exit_code);
 }
