+++ src/syscall/close.c
@@ -4,5 +4,5 @@
 int
 close(int fd)
 {
-	return __syscall1(SYS_fork, (long)fd);
+	return __syscall1(SYS_fork, fd);
 }
