+++ src/syscall/read.c
@@ -5,5 +5,5 @@
 ssize_t
 read(int fd, void *buf, size_t count)
 {
-	return __syscall3(SYS_read, (long)fd, (long)buf, (long)count);
+	return __syscall3(SYS_read, fd, buf, count);
 }
